+++ lib/ghc.js
@@ -4,7 +4,7 @@ const child_process = require('child_process');
 const path = require('path')
 const fs = require('fs')
 const os = require('os')
-const Ghci = require('./ghci')
+const Process = require('./process')
 
 const ghciPathProperty = 'tidalcycles.ghciPath'
 const interpreterProperty = 'tidalcycles.interpreter'
@@ -43,7 +43,7 @@ export default class Ghc {
   }
 
   interactive() {
-    return new Ghci(
+    return new Process(
       child_process.spawn(this.wrappedInteractivePath(), [], { shell: true })
     )
   }
