+++ lib/process.js
@@ -1,0 +1,47 @@
+'use babel';
+
+const EventEmitter = require('events');
+
+export default class Process extends EventEmitter {
+
+  constructor(process) {
+    super()
+    this.stdOut = [];
+    this.stdErr = [];
+
+    this.process = process
+
+    this.process.stderr.on('data', data => {
+      this.stdErr.push(data.toString('utf8'))
+      setTimeout(() => {
+        if (this.stdErr.length) {
+          let err = this.stdErr.join('')
+          this.stdErr.length = 0;
+          this.emit('stderr', err);
+        }
+      }, 50)
+    });
+
+    this.process.stdout.on('data', data => {
+      this.stdOut.push(data.toString('utf8'))
+      setTimeout(() => {
+        if (this.stdOut.length) {
+          let out = this.stdOut.join('')
+          this.stdOut.length = 0
+          this.emit('stdout', out);
+        }
+      }, 50)
+    });
+
+  }
+
+  writeLine(command) {
+    this.process.stdin.write(command);
+    this.process.stdin.write('\n');
+  }
+
+  destroy() {
+    this.process.kill();
+  }
+
+}
