+++ lib/repl.js
@@ -4,14 +4,9 @@ export default class REPL {
 
   ghci = null
   consoleView = null
-  stdErr = null
-  stdOut = null
-  stdTimer = 0
 
   constructor(consoleView, ghc, bootTidal, status, editor) {
     this.consoleView = consoleView;
-    this.stdErr = []
-    this.stdOut = []
     this.ghc = ghc
     this.bootTidal = bootTidal
     this.status = status
@@ -28,64 +23,10 @@ export default class REPL {
     this.status.reset()
 
     this.ghci = this.ghc.interactive()
-      .on('stderr', data => { this.processStdErr(data) })
-      .on('stdout', data => { this.processStdOut(data) })
+      .on('stderr', data => this.consoleView.logPromptErr(this.cleanStdErr(data)))
+      .on('stdout', data => this.consoleView.logPromptOut(this.cleanStdOut(data)))
 
     this.initTidal();
-  }
-
-  hush() {
-    this.tidalSendExpression('hush');
-  }
-
-  processStdOut(data) {
-    this.stdOut.push(data.toString('utf8'))
-    this.processStd()
-  }
-
-  processStdErr(data) {
-    this.stdErr.push(data.toString('utf8'))
-    this.processStd()
-  }
-
-  processStd() {
-    clearTimeout(this.stdTimer)
-    // defers the handler of stdOut/stdErr data
-    // by some arbitrary ammount of time (50ms)
-    // to get the buffer filled completly
-    this.stdTimer = setTimeout(() => {
-      this.flushStdErr()
-      this.flushStdOut()
-    }, 50);
-  }
-
-  flushStdOut() {
-    if (this.stdOut.length) {
-      let output = this.stdOut.join('')
-        .trim()
-        .replace(/tidal>.*Prelude>/g, "")
-        .replace(/tidal>/g, "")
-        .replace(/Prelude>/g, "")
-        .replace(/Prelude.*\|/g, "")
-        .replace(/GHCi.*help/g, "")
-
-      this.consoleView.logPromptOut(output);
-      this.stdOut.length = 0
-    }
-  }
-
-  flushStdErr() {
-    if (this.stdErr.length) {
-      let output = this.stdErr.join('')
-        .trim()
-        .replace(/<interactive>.*error:/g, "")
-        .replace(/ \(bound at.*/g, "")
-
-      this.consoleView.logPromptErr(output);
-      this.stdErr.length = 0
-      // dont care about stdOut if there are errors
-      this.stdOut.length = 0
-    }
   }
 
   initTidal() {
@@ -101,8 +42,26 @@ export default class REPL {
         : this.tidalSendExpression(block)
       )
     }
+  }
 
-    this.tidalSendLine('putStrLn $ "[TidalCycles v" ++ tidal_version ++ "]"')
+  hush() {
+    this.tidalSendExpression('hush');
+  }
+
+  cleanStdOut(stdout) {
+    return stdout
+      .trim()
+      .replace(/tidal>.*Prelude>/g, "")
+      .replace(/tidal>/g, "")
+      .replace(/Prelude>/g, "")
+      .replace(/Prelude.*\|/g, "")
+      .replace(/GHCi.*help/g, "")
+  }
+
+  cleanStdErr(stderr) {
+    return stderr
+      .replace(/<interactive>.*error:/g, "")
+      .replace(/ \(bound at.*/g, "")
   }
 
   tidalSendExpression(expression) {
