+++ lib/tidal-listener-repl.js
@@ -3,6 +3,7 @@
 const dgram = require('dgram');
 const osc = require('osc-min');
 const child_process = require('child_process');
+const Process = require('child_process');
 
 export default class Repl {
 
@@ -25,27 +26,9 @@ export default class Repl {
   start() {
     this.status.reset()
 
-    this.listener = child_process.spawn('tidal-listener', []);
-    this.listener.stderr.on('data', data => {
-      this.stdErr.push(data.toString('utf8'))
-      setTimeout(() => {
-        if (this.stdErr.length) {
-          let err = this.stdErr.join('')
-          this.stdErr.length = 0;
-          this.consoleView.logStderr(err)
-        }
-      }, 50)
-    });
-    this.listener.stdout.on('data', data => {
-      this.stdOut.push(data.toString('utf8'))
-      setTimeout(() => {
-        if (this.stdOut.length) {
-          let log = this.stdOut.join('')
-          this.stdOut.length = 0
-          this.consoleView.logStdlog(log)
-        }
-      }, 50)
-    });
+    this.listener = new Process(child_process.spawn('tidal-listener', []));
+    this.listener.on('stderr', data => this.consoleView.logStderr(data));
+    this.listener.on('stdout', data => this.consoleView.logStdlog(data))
 
     this.udp = dgram.createSocket('udp4');
 
@@ -101,7 +84,7 @@ export default class Repl {
   eval(evalType, copy) {
     if (!this.editor.isTidal()) return;
 
-    if (!this.listener || this.listener.killed) this.start();
+    if (!this.listener) this.start();
     this.editor.currentEvaluations(evalType)
       .filter(eval => eval.expression && eval.range)
       .forEach(eval => {
@@ -144,11 +127,13 @@ export default class Repl {
 
   destroy() {
     if (this.listener) {
-      this.listener.kill();
+      this.listener.destroy();
     }
     if (this.udp) {
       this.udp.close();
     }
+    this.listener = undefined;
+    this.udp = undefined;
   }
 
 }
